+++ cmd/lakectl/cmd/branch_protect.go
@@ -1,6 +1,8 @@
 package cmd
 
 import (
+	"net/http"
+
 	"github.com/spf13/cobra"
 	"github.com/treeverse/lakefs/pkg/api"
 )
@@ -18,13 +20,13 @@ var branchProtectCmd = &cobra.Command{
 var branchProtectListCmd = &cobra.Command{
 	Use:     "list <repo uri>",
 	Short:   "List all branch protection rules",
-	Example: "lakectl list lakefs://<repository>",
+	Example: "lakectl branch-protect list lakefs://<repository>",
 	Args:    cobra.ExactArgs(1),
 	Run: func(cmd *cobra.Command, args []string) {
 		client := getClient()
 		u := MustParseRepoURI("repository", args[0])
 		resp, err := client.GetBranchProtectionRulesWithResponse(cmd.Context(), u.Repository)
-		DieOnResponseError(resp, err)
+		DieOnErrorOrUnexpectedStatusCode(resp, err, http.StatusOK)
 		patterns := make([][]interface{}, len(*resp.JSON200))
 		for i, rule := range *resp.JSON200 {
 			patterns[i] = []interface{}{rule.Pattern}
@@ -40,7 +42,7 @@ var branchProtectAddCmd = &cobra.Command{
 	Use:     "add <repo uri> <pattern>",
 	Short:   "Add a branch protection rule",
 	Long:    "Add a branch protection rule for a given branch name pattern",
-	Example: "lakectl add lakefs://<repository> 'stable_*'",
+	Example: "lakectl branch-protect add lakefs://<repository> 'stable_*'",
 	Args:    cobra.ExactArgs(branchProtectAddCmdArgs),
 	Run: func(cmd *cobra.Command, args []string) {
 		client := getClient()
@@ -48,7 +50,7 @@ var branchProtectAddCmd = &cobra.Command{
 		resp, err := client.CreateBranchProtectionRuleWithResponse(cmd.Context(), u.Repository, api.CreateBranchProtectionRuleJSONRequestBody{
 			Pattern: args[1],
 		})
-		DieOnResponseError(resp, err)
+		DieOnErrorOrUnexpectedStatusCode(resp, err, http.StatusNoContent)
 	},
 }
 
@@ -56,7 +58,7 @@ var branchProtectDeleteCmd = &cobra.Command{
 	Use:     "delete <repo uri> <pattern>",
 	Short:   "Delete a branch protection rule",
 	Long:    "Delete a branch protection rule for a given branch name pattern",
-	Example: "lakectl delete lakefs://<repository> stable_*",
+	Example: "lakectl branch-protect delete lakefs://<repository> stable_*",
 	Args:    cobra.ExactArgs(branchProtectDeleteCmdArgs),
 	Run: func(cmd *cobra.Command, args []string) {
 		client := getClient()
@@ -64,7 +66,7 @@ var branchProtectDeleteCmd = &cobra.Command{
 		resp, err := client.DeleteBranchProtectionRuleWithResponse(cmd.Context(), u.Repository, api.DeleteBranchProtectionRuleJSONRequestBody{
 			Pattern: args[1],
 		})
-		DieOnResponseError(resp, err)
+		DieOnErrorOrUnexpectedStatusCode(resp, err, http.StatusNoContent)
 	},
 }