+++ cmd/lakectl/cmd/common_helpers.go
@@ -2,22 +2,25 @@ package cmd
 
 import (
 	"bytes"
+	"encoding/base64"
 	"encoding/json"
 	"errors"
 	"fmt"
 	"io"
 	"net/http"
 	"os"
+	"reflect"
+	"regexp"
+	"strconv"
 	"strings"
 	"text/template"
 	"time"
-
-	"github.com/treeverse/lakefs/pkg/uri"
 
 	"github.com/jedib0t/go-pretty/v6/table"
 	"github.com/jedib0t/go-pretty/v6/text"
 	"github.com/treeverse/lakefs/pkg/api"
 	"github.com/treeverse/lakefs/pkg/api/helpers"
+	"github.com/treeverse/lakefs/pkg/uri"
 	"golang.org/x/term"
 )
 
@@ -26,6 +29,8 @@ var noColorRequested = false
 
 // ErrInvalidValueInList is an error returned when a parameter of type list contains an empty string
 var ErrInvalidValueInList = errors.New("empty string in list")
+
+var accessKeyRegexp = regexp.MustCompile(`^AKIA[I|J][A-Z0-9]{14}Q$`)
 
 const (
 	LakectlInteractive        = "LAKECTL_INTERACTIVE"
@@ -202,6 +207,31 @@ func DieOnResponseError(response interface{}, err error) {
 		DieErr(retrievedErr)
 	}
 }
+func DieOnErrorOrUnexpectedStatusCode(response interface{}, err error, expectedStatusCode int) {
+	DieOnResponseError(response, err)
+	var statusCode int
+	if httpResponse, ok := response.(*http.Response); ok {
+		statusCode = httpResponse.StatusCode
+	} else {
+		r := reflect.Indirect(reflect.ValueOf(response))
+		f := r.FieldByName("HTTPResponse")
+		httpResponse, _ := f.Interface().(*http.Response)
+		if httpResponse != nil {
+			statusCode = httpResponse.StatusCode
+		}
+	}
+
+	if statusCode == 0 {
+		Die("could not get status code from response", 1)
+	}
+	if statusCode != expectedStatusCode {
+		// redirected to not found page
+		if statusCode == http.StatusFound {
+			Die("got not-found error, probably wrong endpoint url", 1)
+		}
+		Die("got unexpected status code: "+strconv.Itoa(statusCode)+", expected: "+strconv.Itoa(expectedStatusCode), 1)
+	}
+}
 
 func DieOnHTTPError(httpResponse *http.Response) {
 	err := helpers.HTTPResponseAsError(httpResponse)
@@ -266,4 +296,17 @@ func MustParsePathURI(name, s string) *uri.URI {
 		DieFmt("Invalid '%s': %s", name, uri.ErrInvalidPathURI)
 	}
 	return u
+}
+
+func IsValidAccessKeyID(accessKeyID string) bool {
+	return accessKeyRegexp.MatchString(accessKeyID)
+}
+
+func IsValidSecretAccessKey(secretAccessKey string) bool {
+	return IsBase64(secretAccessKey) && len(secretAccessKey) == 40
+}
+
+func IsBase64(s string) bool {
+	_, err := base64.StdEncoding.DecodeString(s)
+	return err == nil
 }