+++ pkg/config/config.go
@@ -108,9 +108,11 @@ func NewConfig() (*Config, error) {
 const (
 	ListenAddressKey = "listen_address"
 
-	LoggingFormatKey = "logging.format"
-	LoggingLevelKey  = "logging.level"
-	LoggingOutputKey = "logging.output"
+	LoggingFormatKey        = "logging.format"
+	LoggingLevelKey         = "logging.level"
+	LoggingOutputKey        = "logging.output"
+	LoggingFileMaxSizeMBKey = "logging.file_max_size_mb"
+	LoggingFilesKeepKey     = "logging.files_keep"
 
 	ActionsEnabledKey = "actions.enabled"
 
@@ -121,6 +123,7 @@ const (
 
 	BlockstoreTypeKey                    = "blockstore.type"
 	BlockstoreLocalPathKey               = "blockstore.local.path"
+	BlockstoreDefaultNamespacePrefixKey  = "blockstore.default_namespace_prefix"
 	BlockstoreS3RegionKey                = "blockstore.s3.region"
 	BlockstoreS3StreamingChunkSizeKey    = "blockstore.s3.streaming_chunk_size"
 	BlockstoreS3StreamingChunkTimeoutKey = "blockstore.s3.streaming_chunk_timeout"
@@ -165,6 +168,7 @@ func setDefaults() {
 	viper.SetDefault(LoggingFormatKey, DefaultLoggingFormat)
 	viper.SetDefault(LoggingLevelKey, DefaultLoggingLevel)
 	viper.SetDefault(LoggingOutputKey, DefaultLoggingOutput)
+	viper.SetDefault(LoggingFilesKeepKey, DefaultLoggingFilesKeepKey)
 
 	viper.SetDefault(ActionsEnabledKey, DefaultActionsEnabled)
 
@@ -302,6 +306,10 @@ func (c *Config) GetAwsConfig() *aws.Config {
 
 func (c *Config) GetBlockstoreType() string {
 	return c.values.Blockstore.Type
+}
+
+func (c *Config) GetBlockstoreDefaultNamespacePrefix() string {
+	return c.values.Blockstore.DefaultNamespacePrefix
 }
 
 func (c *Config) GetBlockAdapterS3Params() (blockparams.S3, error) {